+++ ProcessLayer/Entities/Kiosks/OuterPortRequest.cs
@@ -4,7 +4,7 @@ namespace ProcessLayer.Entities.Kiosk
 {
     public class OuterPortRequest : KioskBase2
     {
-        public byte? LocationID { get; set; }    
+        public int? LocationID { get; set; }    
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
         public bool? IsHighRisk { get; set; } = false;
