+++ ProcessLayer/Entities/Lookups/Deduction.cs
@@ -2,7 +2,7 @@
 {
     public class Deduction : Lookup<int>
     {
-        public byte? WhenToDeduct { get; set; }
+        public int? WhenToDeduct { get; set; }
         public bool? ComputedThruSalary { get; set; }
         public bool? GovernmentDeduction { get; set; }
         public bool? AutoCompute { get; set; }
