+++ ProcessLayer/Entities/Lookups/KioskApprovers.cs
@@ -5,7 +5,7 @@
         public int ID { get; set; }
         public int? DepartmentID { get; set; }
         public long? ApproverID { get; set; }
-        public byte? Sequence { get; set; }
+        public int? Sequence { get; set; }
         public bool Deleted { get; set; } = false;
 
         public Personnel _Personnel { get; set; }
