+++ ProcessLayer/Entities/Lookups/TaxTable.cs
@@ -10,7 +10,7 @@ namespace ProcessLayer.Entities
         public decimal? FixedTax { get; set; }
         public decimal? AdditionalTax { get; set; }
         public decimal? ExcessOver { get; set; }
-        public byte? TaxScheduleID { get; set; }
+        public int? TaxScheduleID { get; set; }
         public DateTime? EffectiveStartDate { get; set; }
         public DateTime? EffectiveEndDate { get; set; }
